stock-checker/pyproject.toml
[project]
name = "stock-checker"
version = "0.1.0"
description = "Product availability monitor for retail stores"
requires-python = ">=3.11"
dependencies = [
    "httpx",
    "pyyaml",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
